Add files via upload
The objective of the proposed system is to develop a software that retrieved data from the Phone data base i.e., device_features.csv whereas the data set is already pre-processed. The aim of the project is to analyse, and visually represent the different features of the devices. The system provides the different characterises of devices based on the user defined features in the phase A, in the Phase B the system implemented the panda data structure to provide the collective characterises of the devices such as top 5 regions, average price of devices, average mass for each manufacturer and the correlation between the features of the device. The last phase of the system consists of the data visualization techniques which include the pie chart to represent the proportion of RAM types for devices, visually compare the number of devices for each USB connector type, represents the monthly average price trends (in GBP) for devices released in each year from 2020 to 2023 and represent the relation between the display diagonal and battery capacity of the devices.