Add files via upload
The data file, device_features.csv, contains 48 columns. Each row in the file represents a single record
for a device. The data set contains complete data for all columns for each record in the file. This means
that there are no missing values. Some columns may contain multiple values within the same column.
It is recommended that you familiarise yourself with the content of the data file before attempting
the remainder of this assessment.